메모리 최적화 목표: vLLM(GPU5, ~9.6GB)과 같은 물리 GPU에서 돌아가는 ConceptGraph 생성 프로세스의 GPU 사용량을 ~16GB에서 ~8GB로 줄여 여유 마진을 확보하되, 정확도/속도는 유지.
원인: run_mapping_for_scene()에서 YOLO-World(fp16)·SAM(fp16)·CLIP ViT-H-14(fp16)·DINOv3 vith16plus가 프로세스 전체 수명 동안 GPU에 동시 상주하는데, 이 중 DINOv3(3.36GB)만 과거에 .half()(fp16) 캐스팅 시 NaN이 나서 fp32로 남아있었음. 여기에 discover_scene_vocabulary의 SAM "segment everything" 호출만 fp32로 빠져있었고, torch.cuda.empty_cache()가 코드 전체에 한 번도 없어서 allocator가 예약한 메모리가 계속 쌓이는 구조였음.

변경 내용:

concept-graphs/conceptgraph/slam/rerun_realtime_mapping.py:217-227 — DINOv3를 fp32 대신 bfloat16으로 로드(.to(torch.bfloat16)). fp16이 NaN을 낸 건 exponent 범위가 좁아서였는데, bf16은 fp32와 같은 exponent 범위를 가져 같은 overflow 없이 메모리를 절반(3.36GB→~1.7GB)으로 줄일 것으로 기대. main()의 before/after 루프 사이, 그리고 vocabulary discovery 직후에 torch.cuda.empty_cache() 추가.
concept-graphs/conceptgraph/utils/model_utils.py의 compute_dinov3_dense_features() — 입력을 bf16으로 캐스팅해 모델과 dtype 맞추고, forward 결과는 다시 .float()로 되돌려 이후 interpolate/정규화/저장 로직(전부 fp32 가정)은 그대로 유지.
concept-graphs/conceptgraph/utils/general_utils.py:478 — vocabulary discovery의 SAM 호출에 메인 루프와 동일하게 quantize=16 추가(정밀도 통일, 피크 완화).
scripts/construct-concept-graphs.sh — PYTORCH_CUDA_ALLOC_CONF=expandable_segments:True 환경변수 추가(allocator 파편화 완화, 결과/속도에 영향 없음).

+++ concept-graphs/conceptgraph/utils/model_utils.py
@@ -94,9 +94,10 @@ def compute_dinov3_dense_features(image_rgb, dinov3_model, device):
     Returns (D, H, W) L2-normalized torch tensor on `device`.
     '''
     h, w = image_rgb.shape[:2]
-    x = dinov3_preprocess(image_rgb).unsqueeze(0).to(device)
+    # dinov3_model runs in bfloat16 (see rerun_realtime_mapping.py) -- match input dtype.
+    x = dinov3_preprocess(image_rgb).unsqueeze(0).to(device, dtype=torch.bfloat16)
     with torch.no_grad():
-        patch_tokens = dinov3_model.forward_features(x)["x_norm_patchtokens"]
+        patch_tokens = dinov3_model.forward_features(x)["x_norm_patchtokens"].float()
     n_side = int(patch_tokens.shape[1] ** 0.5)
     feat_grid = patch_tokens.reshape(1, n_side, n_side, -1).permute(0, 3, 1, 2)
     dense = F.interpolate(feat_grid, size=(h, w), mode="bilinear", align_corners=False)[0]
